+++ src/index.ts
@@ -1,4 +1,4 @@
-export { createProxyHandlers } from './create-proxy-handlers';
+export { createProxyBridge } from './create-proxy-bridge';
 export { sanitizeTokenResponse } from './response/sanitize/sanitize-token-response.util';
 export type {
   AuthCookieConfig,
@@ -7,14 +7,18 @@ export type {
   AuthTokens,
   BackendUrlContext,
   HttpMethod,
+  ProxyBridge,
+  ProxyBridgeConfig,
   ProxyRefreshConfig,
   ProxyAuthConfig,
   ProxyConfig,
   ProxyContext,
   ProxyCookiesConfig,
   ProxyHandler,
+  ProxyHandlers,
   ProxyRequest,
   RefreshRequestContext,
   SanitizeTokenResponseContext,
   SanitizeTokenResponseMode,
+  ServerProxyFetch,
 } from './types';